Bug 1014647 - Fix login session to last for 1 year
The default duration for the login session within profile2 is set to 365
seconds. This is because we poorly documented that config option, so it
sounds like you configure it as days, but it's actually configured as
seconds.

Since we have a default of 31536000000 seconds within webmaker-auth, we
can just remove the configuration option to have a consistent login
experience across our sites.